+++ test/UnitTests.js
@@ -1,68 +1,68 @@
 'use strict'
 
-var BitX = require('../lib/BitX')
-var FakeRequest = require('./helpers/FakeRequest')
-var https = require('https')
-var sinon = require('sinon')
-var tap = require('tap')
+const Luno = require('../lib/luno')
+const FakeRequest = require('./helpers/FakeRequest')
+const https = require('https')
+const sinon = require('sinon')
+const tap = require('tap')
 
-tap.test('BitX constructor', {autoend: true}, function (t) {
+tap.test('Luno constructor', {autoend: true}, function (t) {
   t.test('should create a new instance', function (tt) {
-    var bitx = new BitX()
-    tt.type(bitx, 'BitX')
+    const luno = new Luno()
+    tt.type(luno, 'Luno')
     tt.end()
   })
 
   tap.test('should create a new instance without the new keyword', function (tt) {
-    var bitx = BitX()
-    tt.type(bitx, 'BitX')
+    const luno = Luno()
+    tt.type(luno, 'Luno')
     tt.end()
   })
 
   tap.test('should have default options', function (tt) {
-    var bitx = new BitX()
-    tt.equal(bitx.hostname, 'api.luno.com')
-    tt.equal(bitx.port, 443)
-    tt.equal(bitx.pair, 'XBTZAR')
+    const luno = new Luno()
+    tt.equal(luno.hostname, 'api.luno.com')
+    tt.equal(luno.port, 443)
+    tt.equal(luno.pair, 'XBTZAR')
     tt.end()
   })
 
   tap.test('should accept options', function (tt) {
-    var options = {
+    const options = {
       hostname: 'localhost',
       port: 8000,
       pair: 'XBTUSD'
     }
-    var bitx = new BitX(options)
-    tt.equal(bitx.hostname, options.hostname)
-    tt.equal(bitx.port, options.port)
-    tt.equal(bitx.pair, options.pair)
+    const luno = new Luno(options)
+    tt.equal(luno.hostname, options.hostname)
+    tt.equal(luno.port, options.port)
+    tt.equal(luno.pair, options.pair)
     tt.end()
   })
 
   tap.test('should accept auth and options', function (tt) {
-    var keyId = 'cnz2yjswbv3jd'
-    var keySecret = '0hydMZDb9HRR3Qq-iqALwZtXLkbLR4fWxtDZvkB9h4I'
-    var options = {
+    const keyId = 'cnz2yjswbv3jd'
+    const keySecret = '0hydMZDb9HRR3Qq-iqALwZtXLkbLR4fWxtDZvkB9h4I'
+    const options = {
       hostname: 'localhost',
       port: 8000,
       pair: 'XBTUSD'
     }
-    var bitx = new BitX(keyId, keySecret, options)
-    tt.equal(bitx.hostname, options.hostname)
-    tt.equal(bitx.port, options.port)
-    tt.equal(bitx.pair, options.pair)
-    tt.equal(bitx.auth, keyId + ':' + keySecret)
+    const luno = new Luno(keyId, keySecret, options)
+    tt.equal(luno.hostname, options.hostname)
+    tt.equal(luno.port, options.port)
+    tt.equal(luno.pair, options.pair)
+    tt.equal(luno.auth, keyId + ':' + keySecret)
     tt.end()
   })
 })
 
 tap.test('Internal', {autoend: true}, function (t) {
-  var keyId = '12345'
-  var keySecret = '0000000000000000'
-  var path = '/api/1/test'
+  const keyId = '12345'
+  const keySecret = '0000000000000000'
+  const path = '/api/1/test'
 
-  var expectedOptions = {
+  const expectedOptions = {
     headers: {
       'Accept': 'application/json',
       'Accept-Charset': 'utf-8'
@@ -73,9 +73,9 @@ tap.test('Internal', {autoend: true}, function (t) {
     auth: keyId + ':' + keySecret
   }
 
-  var bitx = new BitX(keyId, keySecret)
-  var request = sinon.stub(https, 'request')
-  var stub = request.withArgs(sinon.match(expectedOptions))
+  const luno = new Luno(keyId, keySecret)
+  const request = sinon.stub(https, 'request')
+  const stub = request.withArgs(sinon.match(expectedOptions))
 
   t.afterEach(function () {
     stub.reset()
@@ -86,9 +86,9 @@ tap.test('Internal', {autoend: true}, function (t) {
   })
 
   t.test('_request should return the expected result', function (tt) {
-    var expectedResult = {success: true}
+    const expectedResult = {success: true}
     stub.returns(new FakeRequest(expectedResult))
-    bitx._request('GET', path, null, function (err, result) {
+    luno._request('GET', path, null, function (err, result) {
       tt.error(err)
       tt.same(result, expectedResult)
       tt.end()
@@ -97,7 +97,7 @@ tap.test('Internal', {autoend: true}, function (t) {
 
   t.test('_request should return an error if request emits an error', function (tt) {
     stub.returns(new FakeRequest(null, {fail: true}))
-    bitx._request('GET', path, null, function (err) {
+    luno._request('GET', path, null, function (err) {
       tt.type(err, 'Error')
       tt.end()
     })
@@ -105,7 +105,7 @@ tap.test('Internal', {autoend: true}, function (t) {
 
   t.test('_request should return an error if the response is not valid', function (tt) {
     stub.returns(new FakeRequest('invalid', {stringify: false}))
-    bitx._request('GET', path, null, function (err) {
+    luno._request('GET', path, null, function (err) {
       tt.type(err, 'Error')
       tt.end()
     })
@@ -113,7 +113,7 @@ tap.test('Internal', {autoend: true}, function (t) {
 
   t.test('_request should return an error if the response contains an error', function (tt) {
     stub.returns(new FakeRequest({error: true}))
-    bitx._request('GET', path, null, function (err) {
+    luno._request('GET', path, null, function (err) {
       tt.type(err, 'Error')
       tt.end()
     })
@@ -121,7 +121,7 @@ tap.test('Internal', {autoend: true}, function (t) {
 
   t.test('_request should return an error if the response is unauthorized', function (tt) {
     stub.returns(new FakeRequest(null, {statusCode: 401}))
-    bitx._request('GET', path, null, function (err) {
+    luno._request('GET', path, null, function (err) {
       tt.type(err, 'Error')
       tt.end()
     })
@@ -129,12 +129,12 @@ tap.test('Internal', {autoend: true}, function (t) {
 })
 
 tap.test('External', {autoend: true}, function (t) {
-  var bitx = new BitX()
-  var callback = function () {}
-  var mock
+  const luno = new Luno()
+  const callback = function () {}
+  let mock
 
   t.beforeEach(function () {
-    mock = sinon.mock(bitx)
+    mock = sinon.mock(luno)
   })
 
   t.afterEach(function () {
@@ -143,149 +143,149 @@ tap.test('External', {autoend: true}, function (t) {
 
   t.test('getTicker should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/ticker', {pair: 'XBTZAR'}, callback)
-    bitx.getTicker(callback)
+    luno.getTicker(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getTicker should accept a pair option', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/ticker', {pair: 'XBTMYR'}, callback)
-    bitx.getTicker({pair: 'XBTMYR'}, callback)
+    luno.getTicker({pair: 'XBTMYR'}, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getAllTickers should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/tickers', null, callback)
-    bitx.getAllTickers(callback)
+    luno.getAllTickers(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getOrderBook should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/orderbook', {pair: 'XBTZAR'}, callback)
-    bitx.getOrderBook(callback)
+    luno.getOrderBook(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getOrderBook should accept a pair option', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/orderbook', {pair: 'XBTMYR'}, callback)
-    bitx.getOrderBook({pair: 'XBTMYR'}, callback)
+    luno.getOrderBook({pair: 'XBTMYR'}, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getTrades should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/trades', {pair: 'XBTZAR'}, callback)
-    bitx.getTrades(callback)
+    luno.getTrades(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getTrades should accept a pair option', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/trades', {pair: 'XBTMYR'}, callback)
-    bitx.getTrades({pair: 'XBTMYR'}, callback)
+    luno.getTrades({pair: 'XBTMYR'}, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getOrderList should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/listorders', {pair: 'XBTZAR'}, callback)
-    bitx.getOrderList(callback)
+    luno.getOrderList(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getOrderList should accept a pair option', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/listorders', {pair: 'XBTMYR'}, callback)
-    bitx.getOrderList({pair: 'XBTMYR'}, callback)
+    luno.getOrderList({pair: 'XBTMYR'}, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getLimits should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/BTCZAR/getlimits', null, callback)
-    bitx.getLimits(callback)
+    luno.getLimits(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('postBuyOrder should call _request with the correct parameters', function (tt) {
-    var parameters = {
+    const parameters = {
       type: 'BID',
       volume: 9999.99,
       price: 0.001,
       pair: 'XBTZAR'
     }
     mock.expects('_request').once().withArgs('POST', '/api/1/postorder', parameters, callback)
-    bitx.postBuyOrder(parameters.volume, parameters.price, callback)
+    luno.postBuyOrder(parameters.volume, parameters.price, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('postMarketBuyOrder should call _request with the correct parameters', function (tt) {
-    var parameters = {
+    const parameters = {
       type: 'BUY',
       counter_volume: 9999.99,
       pair: 'XBTZAR'
     }
     mock.expects('_request').once().withArgs('POST', '/api/1/marketorder', parameters, callback)
-    bitx.postMarketBuyOrder(parameters.counter_volume, callback)
+    luno.postMarketBuyOrder(parameters.counter_volume, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('postSellOrder should call _request with the correct parameters', function (tt) {
-    var parameters = {
+    const parameters = {
       type: 'ASK',
       volume: 0.001,
       price: 9999.99,
       pair: 'XBTZAR'
     }
     mock.expects('_request').once().withArgs('POST', '/api/1/postorder', parameters, callback)
-    bitx.postSellOrder(parameters.volume, parameters.price, callback)
+    luno.postSellOrder(parameters.volume, parameters.price, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('postMarketSellOrder should call _request with the correct parameters', function (tt) {
-    var parameters = {
+    const parameters = {
       type: 'SELL',
       base_volume: 9999.99,
       pair: 'XBTZAR'
     }
     mock.expects('_request').once().withArgs('POST', '/api/1/marketorder', parameters, callback)
-    bitx.postMarketSellOrder(parameters.base_volume, callback)
+    luno.postMarketSellOrder(parameters.base_volume, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('stopOrder should call _request with the correct parameters', function (tt) {
-    var body = {order_id: 'BXMC2CJ7HNB88U4'}
+    const body = {order_id: 'BXMC2CJ7HNB88U4'}
     mock.expects('_request').once().withArgs('POST', '/api/1/stoporder', body, callback)
-    bitx.stopOrder(body.order_id, callback)
+    luno.stopOrder(body.order_id, callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getBalance should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/balance', null, callback)
-    bitx.getBalance(callback)
+    luno.getBalance(callback)
     mock.verify()
     tt.end()
   })
 
   t.test('getBalance should accept an asset argument and call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/balance', {asset: 'ZAR'}, callback)
-    bitx.getBalance('ZAR', callback)
+    luno.getBalance('ZAR', callback)
     mock.verify()
     tt.end()
   })
 
   t.test('should call _request with the correct parameters', function (tt) {
     mock.expects('_request').once().withArgs('GET', '/api/1/funding_address', {asset: 'XBT'}, callback)
-    bitx.getFundingAddress('XBT', callback)
+    luno.getFundingAddress('XBT', callback)
     mock.verify()
     tt.end()
   })